requirements.txt
langchain
langchain-community
google-generativeai
python-dotenv

